+++ mod-record-specifications-server/src/main/java/org/folio/rspec/controller/handler/HibernateConstraintViolationExceptionHandler.java
@@ -4,6 +4,7 @@ import static org.folio.rspec.controller.handler.ServiceExceptionHandler.fromErr
 import static org.folio.rspec.domain.dto.ErrorCode.DUPLICATE_FIELD_INDICATOR;
 import static org.folio.rspec.domain.dto.ErrorCode.DUPLICATE_FIELD_TAG;
 import static org.folio.rspec.domain.dto.ErrorCode.DUPLICATE_INDICATOR_CODE;
+import static org.folio.rspec.domain.dto.ErrorCode.DUPLICATE_SUBFIELD;
 import static org.folio.rspec.domain.dto.ErrorCode.UNEXPECTED;
 
 import java.util.Map;
@@ -13,6 +14,7 @@ import org.folio.rspec.domain.dto.ErrorCollection;
 import org.folio.rspec.domain.entity.Field;
 import org.folio.rspec.domain.entity.Indicator;
 import org.folio.rspec.domain.entity.IndicatorCode;
+import org.folio.rspec.domain.entity.Subfield;
 import org.folio.rspec.service.i18n.ExtendedTranslationService;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -26,7 +28,8 @@ public class HibernateConstraintViolationExceptionHandler implements ServiceExce
   private static final Map<String, ErrorCode> DB_CONSTRAINTS_MAP = Map.of(
     Field.TAG_UNIQUE_CONSTRAINT, DUPLICATE_FIELD_TAG,
     Indicator.ORDER_UNIQUE_CONSTRAINT, DUPLICATE_FIELD_INDICATOR,
-    IndicatorCode.CODE_UNIQUE_CONSTRAINT, DUPLICATE_INDICATOR_CODE
+    IndicatorCode.CODE_UNIQUE_CONSTRAINT, DUPLICATE_INDICATOR_CODE,
+    Subfield.SUBFIELD_CODE_UNIQUE_CONSTRAINT, DUPLICATE_SUBFIELD
   );
 
   private final ExtendedTranslationService translationService;