+++ mod-record-specifications-server/src/test/java/org/folio/rspec/service/sync/SpecificationSyncServiceTest.java
@@ -3,13 +3,14 @@ package org.folio.rspec.service.sync;
 import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import org.folio.rspec.domain.dto.Scope;
 import org.folio.rspec.domain.entity.Field;
@@ -50,14 +51,14 @@ class SpecificationSyncServiceTest {
     specification.setId(specId);
 
     var fieldsArray = prepareFetchedFields();
+    ArgumentCaptor<Collection<Field>> fieldsCaptor = ArgumentCaptor.captor();
 
     when(metadataService.getSpecificationMetadata(specId)).thenReturn(metadata);
     when(specificationFetcher.fetch(metadata.getSyncUrl())).thenReturn(fieldsArray);
+    doNothing().when(specificationFieldService).syncFields(any(), fieldsCaptor.capture());
 
     specificationSyncService.sync(specification);
 
-    ArgumentCaptor<List<Field>> fieldsCaptor = ArgumentCaptor.captor();
-    verify(specificationFieldService).syncFields(any(), fieldsCaptor.capture());
     verify(metadataService).saveSpecificationMetadata(metadata);
 
     assertThat(fieldsCaptor.getValue())