+++ NEWS.md
@@ -10,6 +10,7 @@
 * Specification Storage: implement endpoints to manage field definitions ([MRSPECS-6](https://folio-org.atlassian.net//browse/MRSPECS-6))
 * Specification Storage: implement sync endpoint to initialize specification with defaults ([MRSPECS-12](https://folio-org.atlassian.net//browse/MRSPECS-12))
 * Specification Storage: implement endpoints to manage indicator/code definitions ([MRSPECS-7](https://folio-org.atlassian.net//browse/MRSPECS-7))
+* Specification Storage: implement GET/POST endpoints to manage subfield definitions ([MRSPECS-8](https://folio-org.atlassian.net//browse/MRSPECS-8))
 #### General
 * Implement build dependants GitHub workflow on PR creation ([MRSPECS-9](https://folio-org.atlassian.net//browse/MRSPECS-9))